src/cloud-functions/query-simulator/requirements.txt
google-cloud-bigquery==3.31.0
google-cloud-storage==3.1.0
google-api-core==2.24.2